+++ src/assets/css/index.css
@@ -8,7 +8,6 @@ html, body, #root {
 
 main {
   color: white;
-  /* text-shadow: 0 1px 0 black;  */
   font-weight: 500;
   height: 100%;
   display: flex;
@@ -18,6 +17,7 @@ main {
 
 .row {
   display: flex;
+  z-index: 10;
 }
 
 .top-row, .bottom-row {
