+++ src/assets/css/wallpaper.css
@@ -1,0 +1,12 @@
+.wallpaper-container {
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  margin: 0;
+  padding: 0;
+  z-index: 1;
+  background-color: transparent;
+  background: cover center no-repeat;
+}
